Disable STAT_MATCHSTARTTIME sync if gameclock counts down
Since the server doesn't include overtime data, we can't rely on the
STAT_MATCHSTARTTIME data when using a game clock that counts down.
Instead, we'll have to wait for a "[X] minutes remaining" message to
sync the game clock. Without this, during overtime, the clock will start
counting up instead of down.